+++ README.md
@@ -17,36 +17,18 @@ Next, use [docker compose](https://docs.docker.com/compose/) to stand up a local
 ### Set up your `.env` file
 Create a file called `.env` in the root folder of the project and copy the contents of `sample.env` into that new file
 
+### Drizzle ORM setup
+To configure Drizzle:
+
+- Run the migrations with `npm run drizzle:migrate`
+- Load the data by running the sql commands in `drizzle-management/load.sql`
+   - This can be done through a db manager, such as pgAdmin or psql
+
 ### Run the project
 Finally, to run the project locally:
 ```
 npm run dev
 ```
-
-### ORM setup
-Two ORMs are configured to work with the api. However, only one should be used at a time. When switching between the two, it is best to wipe the database and run their migrations fresh. Mikro ORM is the default orm. Drizzle must be opted-into with a feature flag.
-
-#### Database setup
-Before switching to a different ORM, wipe the database:
-1. Remove any running containers `docker compose down`
-2. Remove the database volume `rm -rf db-volume` (this will likely also require `sudo`)
-3. Bring the containers back `docker compose up`
-4. Goto; orm-specific setup
-   - [mikro-orm](#mikro-orm-setup) or
-   - [drizzle-orm](#drizzle-orm-setup)
-
-##### Mikro ORM setup
-1. Check Drizzle is not active. In `.env`, make sure `FEATURE_FLAG_DRIZZLE` is unset
-    - ie) `FEATURE_FLAG_DRIZZLE=` 
-1. Run the migrations with `npm run migrate:up`
-   - The migrations will also load the data
-
-##### Drizzle ORM setup
-1. Opt-in to the drizzle feature. In `.env`, set `FEATURE_FLAG_DRIZZLE=1`
-2. Run the migrations with `npm run drizzle:migrate`
-3. Load the data by running the sql commands in `drizzle-management/load.sql`
-   - This can be done through a db manager, such as pgAdmin or psql
-
 ### OpenApi documentation
 Open Api documentation is managed through [redoc](https://redocly.com/docs/redoc/). Changes to the documentation are managed through edits to the `openapi/openapi.yaml` file. The `openapi/index.html` is used to serve the rendered documentation. It is generated through the redocly cli and should not be manually edited.
