+++ package.json
@@ -13,8 +13,6 @@
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
-    "migrate:create": "mikro-orm migration:create",
-    "migrate:up": "mikro-orm migration:up",
     "drizzle:migrate": "ts-node ./drizzle-management/migrate.ts",
     "redoc:build": "redocly build-docs openapi/openapi.yaml --output openapi/index.html",
     "redoc:dev": "redocly preview-docs openapi/openapi.yaml",
@@ -29,11 +27,6 @@
     "generate": "kubb generate"
   },
   "dependencies": {
-    "@mikro-orm/core": "^5.9.0",
-    "@mikro-orm/migrations": "^5.9.0",
-    "@mikro-orm/nestjs": "^5.2.2",
-    "@mikro-orm/postgresql": "^5.9.0",
-    "@mikro-orm/reflection": "^5.9.0",
     "@nestjs/common": "^10.0.0",
     "@nestjs/config": "^3.1.1",
     "@nestjs/core": "^10.0.0",
@@ -54,7 +47,6 @@
     "@kubb/swagger-client": "^1.14.9",
     "@kubb/swagger-ts": "^1.14.9",
     "@kubb/swagger-zod": "^1.14.9",
-    "@mikro-orm/cli": "^5.9.0",
     "@nestjs/cli": "^10.0.0",
     "@nestjs/schematics": "^10.0.0",
     "@nestjs/testing": "^10.0.0",
@@ -84,13 +76,6 @@
     "ts-node": "^10.9.1",
     "tsconfig-paths": "^4.2.0",
     "typescript": "^5.1.3"
-  },
-  "mikro-orm": {
-    "useTsNode": true,
-    "configPaths": [
-      "./src/mikro-orm.config.ts",
-      "./dist/mikro-orm.config.js"
-    ]
   },
   "jest": {
     "moduleFileExtensions": [
