+++ sample.env
@@ -4,4 +4,3 @@ DATABASE_NAME=zoning
 DATABASE_PORT=5432
 DATABASE_HOST=localhost
 STORAGE_URL=
-FEATURE_FLAG_DRIZZLE=
