+++ src/app.module.ts
@@ -1,8 +1,7 @@
 import { Module } from "@nestjs/common";
 import { AppController } from "./app.controller";
 import { AppService } from "./app.service";
-import { ConfigModule, ConfigService } from "@nestjs/config";
-import { MikroOrmModule } from "@mikro-orm/nestjs";
+import { ConfigModule } from "@nestjs/config";
 import { ServeStaticModule } from "@nestjs/serve-static";
 import * as Joi from "joi";
 import { join } from "path";
@@ -31,20 +30,6 @@ import { ZoningDistrictClassModule } from "./zoning-district-class/zoning-distri
         allowUnknown: true,
         abortEarly: true,
       },
-    }),
-    MikroOrmModule.forRootAsync({
-      imports: [ConfigModule],
-      inject: [ConfigService],
-      useFactory: (configService: ConfigService) => ({
-        type: "postgresql",
-        dbName: configService.get("DATABASE_NAME"),
-        user: configService.get("DATABASE_USER"),
-        password: configService.get("DATABASE_PASSWORD"),
-        host: configService.get("DATABASE_HOST"),
-        port: configService.get("DATABASE_PORT"),
-        entities: ["dist/**/*.entity.js"],
-        entitiesTs: ["src/**/*.entity.ts"],
-      }),
     }),
     ServeStaticModule.forRoot({
       rootPath: join(__dirname, "..", "openapi"),
