+++ src/borough/borough.module.ts
@@ -1,14 +1,11 @@
-import { MikroOrmModule } from "@mikro-orm/nestjs";
 import { Module } from "@nestjs/common";
-import { Borough } from "./borough.entity";
 import { BoroughService } from "./borough.service";
 import { BoroughController } from "./borough.controller";
-import { BoroughRepo } from "./borough.repo";
+import { BoroughRepository } from "./borough.repository";
 
 @Module({
-  imports: [MikroOrmModule.forFeature({ entities: [Borough] })],
   exports: [BoroughService],
-  providers: [BoroughService, BoroughRepo],
+  providers: [BoroughService, BoroughRepository],
   controllers: [BoroughController],
 })
 export class BoroughModule {}
