+++ src/borough/borough.repository.ts
@@ -1,4 +1,18 @@
-import { EntityRepository } from "@mikro-orm/postgresql";
-import { Borough } from "./borough.entity";
+import { Inject } from "@nestjs/common";
+import { DB, DbType } from "src/global/providers/db.provider";
+import { DataRetrievalException } from "src/exception";
 
-export class BoroughRepository extends EntityRepository<Borough> {}
+export class BoroughRepository {
+  constructor(
+    @Inject(DB)
+    private readonly db: DbType,
+  ) {}
+
+  async findAll() {
+    try {
+      return await this.db.query.borough.findMany();
+    } catch {
+      throw new DataRetrievalException();
+    }
+  }
+}
