+++ src/land-use/land-use.module.ts
@@ -1,14 +1,11 @@
-import { MikroOrmModule } from "@mikro-orm/nestjs";
 import { Module } from "@nestjs/common";
-import { LandUse } from "./land-use.entity";
 import { LandUseService } from "./land-use.service";
 import { LandUseController } from "./land-use.controller";
-import { LandUseRepo } from "./land-use.repo";
+import { LandUseRepository } from "./land-use.repository";
 
 @Module({
-  imports: [MikroOrmModule.forFeature({ entities: [LandUse] })],
   exports: [LandUseService],
-  providers: [LandUseService, LandUseRepo],
+  providers: [LandUseService, LandUseRepository],
   controllers: [LandUseController],
 })
 export class LandUseModule {}
