--- src/mikro-orm-pgis/geography/index.ts
@@ -1,1 +1,0 @@
-export { MultiPolygonGeogType } from "./MultiPolygon";
