--- src/mikro-orm-pgis/geometry/MultiPolygon.ts
@@ -1,16 +1,0 @@
-import { Type } from "@mikro-orm/core";
-import { MultiPolygon } from "geojson";
-import { SimpleFeature } from "..";
-
-export class MultiPolygonGeomType extends Type<MultiPolygon, string> {
-  constructor(srid = 3857) {
-    super();
-    this.srid = srid;
-  }
-
-  srid: number;
-
-  getColumnType(): string {
-    return `geometry(${SimpleFeature.MultiPolygon}, ${this.srid})`;
-  }
-}
