--- src/mikro-orm-pgis/geometry/index.ts
@@ -1,1 +1,0 @@
-export { MultiPolygonGeomType } from "./MultiPolygon";
