--- src/mikro-orm-pgis/index.ts
@@ -1,4 +1,0 @@
-export { SimpleFeature } from "./types";
-
-export * from "./geography";
-export * from "./geometry";
