--- src/mikro-orm-pgis/types.ts
@@ -1,3 +1,0 @@
-export enum SimpleFeature {
-  MultiPolygon = "MultiPolygon",
-}
