--- src/mikro-orm.config.ts
@@ -1,19 +1,0 @@
-import { TsMorphMetadataProvider } from "@mikro-orm/reflection";
-import { Migrator } from "@mikro-orm/migrations";
-import { defineConfig } from "@mikro-orm/postgresql";
-
-export default defineConfig({
-  dbName: process.env.DATABASE_NAME,
-  type: "postgresql",
-  user: process.env.DATABASE_USER,
-  password: process.env.DATABASE_PASSWORD,
-  host: process.env.DATABASE_HOST,
-  port: parseInt(process.env.DATABASE_PORT as string),
-  entities: ["dist/**/*.entity.js"],
-  entitiesTs: ["src/**/*.entity.ts"],
-  debug: true,
-  metadataProvider: TsMorphMetadataProvider,
-  // @ts-expect-error nestjs adapter option
-  registerRequestContext: false,
-  extensions: [Migrator],
-});
