+++ src/zoning-district-class/zoning-district-class.service.ts
@@ -1,58 +1,35 @@
 import { Inject, Injectable } from "@nestjs/common";
-import { FeatureFlagConfig } from "src/config";
-import { ConfigType } from "@nestjs/config";
 import { ResourceNotFoundException } from "src/exception";
-import { ZoningDistrictClassRepo } from "./zoning-district-class.repo";
+import { ZoningDistrictClassRepository } from "./zoning-district-class.repository";
 
 @Injectable()
 export class ZoningDistrictClassService {
   constructor(
-    @Inject(ZoningDistrictClassRepo)
-    private readonly zoningDistrictClassRepo: ZoningDistrictClassRepo,
-
-    @Inject(FeatureFlagConfig.KEY)
-    private featureFlagConfig: ConfigType<typeof FeatureFlagConfig>,
+    @Inject(ZoningDistrictClassRepository)
+    private readonly zoningDistrictClassRepository: ZoningDistrictClassRepository,
   ) {}
 
   async findAllZoningDistrictClasses() {
-    if (this.featureFlagConfig.useDrizzle) {
-      const zoningDistrictClasses =
-        await this.zoningDistrictClassRepo.findAll();
-      return {
-        zoningDistrictClasses,
-      };
-    } else {
-      throw new Error(
-        "Zoning District Classes route not supported in Mikro ORM",
-      );
-    }
+    const zoningDistrictClasses =
+      await this.zoningDistrictClassRepository.findAll();
+    return {
+      zoningDistrictClasses,
+    };
   }
 
   async findZoningDistrictClassById(id: string) {
-    if (this.featureFlagConfig.useDrizzle) {
-      const result = this.zoningDistrictClassRepo.findById(id);
-      if (result === undefined) throw new ResourceNotFoundException();
+    const result = this.zoningDistrictClassRepository.findById(id);
+    if (result === undefined) throw new ResourceNotFoundException();
 
-      return result;
-    } else {
-      throw new Error(
-        "Zoning District Classes route not supported in Mikro ORM",
-      );
-    }
+    return result;
   }
 
   async findZoningDistrictClassCategoryColors() {
-    if (this.featureFlagConfig.useDrizzle) {
-      const zoningDistrictClassCategoryColors =
-        await this.zoningDistrictClassRepo.findCategoryColors();
+    const zoningDistrictClassCategoryColors =
+      await this.zoningDistrictClassRepository.findCategoryColors();
 
-      return {
-        zoningDistrictClassCategoryColors,
-      };
-    } else {
-      throw new Error(
-        "Zoning District Classes route not supported in Mikro ORM",
-      );
-    }
+    return {
+      zoningDistrictClassCategoryColors,
+    };
   }
 }
