+++ src/zoning-district/zoning-district.service.ts
@@ -1,54 +1,31 @@
-import { InjectRepository } from "@mikro-orm/nestjs";
 import { Inject, Injectable } from "@nestjs/common";
-import { ZoningDistrict } from "./zoning-district.entity";
-import { ZoningDistrictRepository } from "./zoning-district.repository";
-import { FeatureFlagConfig } from "src/config";
-import { ConfigType } from "@nestjs/config";
 import { ResourceNotFoundException } from "src/exception";
-import { ZoningDistrictRepo } from "./zoning-district.repo";
+import { ZoningDistrictRepository } from "./zoning-district.repository";
 
 @Injectable()
 export class ZoningDistrictService {
   constructor(
-    @InjectRepository(ZoningDistrict)
+    @Inject(ZoningDistrictRepository)
     private readonly zoningDistrictRepository: ZoningDistrictRepository,
-
-    @Inject(ZoningDistrictRepo)
-    private readonly zoningDistrictRepo: ZoningDistrictRepo,
-
-    @Inject(FeatureFlagConfig.KEY)
-    private featureFlagConfig: ConfigType<typeof FeatureFlagConfig>,
   ) {}
 
   async findZoningDistrictByUuid(uuid: string) {
-    if (this.featureFlagConfig.useDrizzle) {
-      const zoningDistrict = await this.zoningDistrictRepo.findByUuid(uuid);
-      if (zoningDistrict === undefined) throw new ResourceNotFoundException();
+    const zoningDistrict = await this.zoningDistrictRepository.findByUuid(uuid);
+    if (zoningDistrict === undefined) throw new ResourceNotFoundException();
 
-      return zoningDistrict;
-    } else {
-      return this.zoningDistrictRepository.findOne(uuid, {
-        fields: ["id", "label"],
-      });
-    }
+    return zoningDistrict;
   }
 
   async findClassesByZoningDistrictUuid(id: string) {
-    if (this.featureFlagConfig.useDrizzle) {
-      const zoningDistrictCheck =
-        await this.zoningDistrictRepo.checkZoningDistrictById(id);
-      if (zoningDistrictCheck === undefined)
-        throw new ResourceNotFoundException();
-      const zoningDistrictClasses =
-        await this.zoningDistrictRepo.findClassesByUuid(id);
+    const zoningDistrictCheck =
+      await this.zoningDistrictRepository.checkZoningDistrictById(id);
+    if (zoningDistrictCheck === undefined)
+      throw new ResourceNotFoundException();
+    const zoningDistrictClasses =
+      await this.zoningDistrictRepository.findClassesByUuid(id);
 
-      return {
-        zoningDistrictClasses,
-      };
-    } else {
-      throw new Error(
-        "Zoning District Classes route not supported in Mikro ORM",
-      );
-    }
+    return {
+      zoningDistrictClasses,
+    };
   }
 }
